+++ src/AppBundle/Resources/config/routing/incident.yml
@@ -14,7 +14,7 @@ caldera_cycleways_incident_comment_add:
 
 caldera_cycleways_incident_status_update:
     path:     /{slug}/update-status
-    defaults: { _controller: AppBundle:IncidentStatus:update }
+    defaults: { _controller: AppBundle:Incident/Status:update }
     requirements:
         slug: '(.*)\-([0-9]{1,6})'
     options:
@@ -22,7 +22,7 @@ caldera_cycleways_incident_status_update:
 
 caldera_cycleways_incident_tag_update:
     path:     /{slug}/update-tag
-    defaults: { _controller: AppBundle:Tag:update }
+    defaults: { _controller: AppBundle:Incident/Tag:update }
     requirements:
         slug: '(.*)\-([0-9]{1,6})'
     options:
@@ -30,7 +30,7 @@ caldera_cycleways_incident_tag_update:
 
 caldera_cycleways_incident_issuer_update:
     path:     /{slug}/update-issuer
-    defaults: { _controller: AppBundle:Issuer:update }
+    defaults: { _controller: AppBundle:Incident/Issuer:update }
     requirements:
         slug: '(.*)\-([0-9]{1,6})'
     options:
